requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dataclasses-json==0.5.7
defusedxml==0.7.1
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==15.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
marshmallow==3.18.0
marshmallow-dataclass==8.5.8
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.3.0
urllib3==1.26.12
